Use `Objects.equals` in translating `==`
When translating `==` for reference types, the backend now uses
`Objects.equals` instead of a null check and `Object.equals`.

One motivation is to remove avoidable synthetic branches/instructions
for null data that are intepreted as uncovered code by JaCoCo etc.

It also reduces bytecode size.

We recently started using `Objects.hashCode` for the synthetic
`hashCode` implementation in value classes (PR 11204).

Note that if either operand is potential `Number` or `Character`,
the backend delegates to `BoxesRuntime.equals` which handles Scala
semantic edge cases for boxed primitives. This PR does not change
that at all.