Deprecate `-Xmain-class``
The Main-Class manifest attribute written in direct-to-jar compilation
is only useful when using the legacy `scala` runner script.

scala-cli does not read the attribute, it detects main classes instead.

Running a Scala jar using `java -jar` does not work either because
the Scala library will not be on the classpath; combining `-jar` with
`-classpath` does not work.